+++ WeatherApp/Networking/Models/Weather/CurrentWeather.swift
@@ -1,0 +1,35 @@
+//
+//  CurrentWeather.swift
+//  WeatherApp
+//
+//  Created by mohammad suhail on 7/6/22.
+//
+
+import Foundation
+
+// MARK: - CurrentWeather
+struct CurrentWeather: Codable {
+    let dt: Int
+    let temp, feelsLike: Double
+    let pressure, humidity: Int
+    let dewPoint, uvi: Double
+    let clouds, visibility: Int
+    let windSpeed: Double
+    let windDeg: Int
+    let weather: [WeatherDetails]
+    let windGust, pop: Double?
+    let rain: Rain?
+
+    enum CodingKeys: String, CodingKey {
+        case dt, temp
+        case feelsLike = "feels_like"
+        case pressure, humidity
+        case dewPoint = "dew_point"
+        case uvi, clouds, visibility
+        case windSpeed = "wind_speed"
+        case windDeg = "wind_deg"
+        case weather
+        case windGust = "wind_gust"
+        case pop, rain
+    }
+}
