+++ WeatherApp/Networking/Models/Weather/FeelsLike.swift
@@ -1,0 +1,13 @@
+//
+//  FeelsLike.swift
+//  WeatherApp
+//
+//  Created by mohammad suhail on 7/6/22.
+//
+
+import Foundation
+
+// MARK: - FeelsLike
+struct FeelsLike: Codable {
+    let day, night, eve, morn: Double
+}
