+++ WeatherApp/Networking/Models/Weather/Weather.swift
@@ -21,70 +21,3 @@ struct Weather: Codable {
         case current, hourly
     }
 }
-
-// MARK: - Current
-struct CurrentWeather: Codable {
-    let dt: Int
-    let sunrise, sunset: Int?
-    let temp, feelsLike: Double
-    let pressure, humidity: Int
-    let dewPoint, uvi: Double
-    let clouds, visibility: Int
-    let windSpeed: Double
-    let windDeg: Int
-    let weather: [WeatherDetails]
-    let windGust, pop: Double?
-    let rain: Rain?
-
-    enum CodingKeys: String, CodingKey {
-        case dt, sunrise, sunset, temp
-        case feelsLike = "feels_like"
-        case pressure, humidity
-        case dewPoint = "dew_point"
-        case uvi, clouds, visibility
-        case windSpeed = "wind_speed"
-        case windDeg = "wind_deg"
-        case weather
-        case windGust = "wind_gust"
-        case pop, rain
-    }
-}
-
-// MARK: - Rain
-struct Rain: Codable {
-    let the1H: Double
-
-    enum CodingKeys: String, CodingKey {
-        case the1H = "1h"
-    }
-}
-
-// MARK: - WeatherElement
-struct WeatherDetails: Codable {
-    let id: Int
-    let main: Main
-    let weatherDescription, icon: String
-
-    enum CodingKeys: String, CodingKey {
-        case id, main
-        case weatherDescription = "description"
-        case icon
-    }
-}
-
-enum Main: String, Codable {
-    case clear = "Clear"
-    case clouds = "Clouds"
-    case rain = "Rain"
-}
-
-// MARK: - FeelsLike
-struct FeelsLike: Codable {
-    let day, night, eve, morn: Double
-}
-
-// MARK: - Temp
-struct Temp: Codable {
-    let day, min, max, night: Double
-    let eve, morn: Double
-}
